pyproject.toml
[project]
name = "data-designer-declarative-columns"
version = "1.0.0"
description = "Data Designer utility for loading column configurations from YAML files"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Maxim Salnikov", email = "salnikov@gmail.com" }
]
keywords = ["data-designer", "yaml", "declarative", "synthetic-data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "data-designer",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/webmaxru/data-designer-declarative-columns"
Repository = "https://github.com/webmaxru/data-designer-declarative-columns"
Issues = "https://github.com/webmaxru/data-designer-declarative-columns/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_designer_declarative_columns"]
